Fix extra_space method
