+++ bin/main.rb
@@ -1,0 +1,3 @@
+require './lib/lint.rb'
+
+document = "./examples/bubble_sort.rb"
