+++ examples/bubble_sort.rb
@@ -1,6 +1,6 @@
 def bubble_sort(array)
   array.length.times do
-    sorted = true  
+    sorted = true
     (array.length - 1).times do |i|
       if array[i] > array[i + 1]
         array[i], array[i + 1] = array[i + 1], array[i]
