+++ lib/parameters.rb
@@ -1,0 +1,8 @@
+class Parameters
+  attr_reader :max_length, :space_indentations
+
+  def intialize
+    @max_length = 10
+    @space_indentations = 2
+  end
+end
